presets/madhu.zincp.zsh: my idiosyncratic prompt in its own segment
* presets/madhu.zincp.zsh: author's indiosyncratic prompt, hardocded
colours for a black on white xterm, default renders without any
fonts. uses zsh-async to run vcs_info.  Had to split segment into
presets/zincs_shlvl/* to avoid surprises (othervice vcs_info is not
always autoloaded in the subprocess)
The segment is currently named "shlvl" for no good reason, and
is a candidate for renaming.

* segments/zincs_shlvl:
* segments/zincs_shlvl_async:
* segments/zincs_shlvl_async_return:
* segments/zincs_shlvl_async_started:
* segments/zincs_shlvl_default_opts: